refactor(tests): Rework Compiler Tests (#7)
Refactor of these tests goes along with refactor of whole system, so
they wont work on their own. Aim was to have cleaner tests that can be
used as example.

This is last of tests that needed updating, following this update, in
TDD fashion I'll update parts of actual package to reflect these changes